Switched to OpenMP.
64-bit version does not link. Something about virtual void I think.

git-svn-id: svn://136.177.114.72/svn_GW/phastpp/branches/phastpp_class@6889 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/phast/IPhreeqcPhast/IPhreeqcPhastLib.cpp
@@ -30,7 +30,7 @@ IPhreeqcPhastLib::CreateIPhreeqcPhast(void)
 		{
 			IPhreeqcPhastPtr = new IPhreeqcPhast;
 		}
-		n = IPhreeqcPhastPtr->Index;
+		n = (int) IPhreeqcPhastPtr->Index;
 	}
 	catch(...)
 	{
